Creating the Task resource with generators, associating with users
Now let’s build the core of the Taskify app: Tasks!

We already have the users and now, we need to allow users to 
create/read/update/delete tasks! 

https://en.wikipedia.org/wiki/Create,_read,_update_and_delete


For this, let’s use a phoenix generator for HTML resources.
Our goal is to build a list of tasks
Each task will have a name to identify each task, an optional description and
a user associated with it!

A user can have many tasks but a task belongs only to a unique user, so, 
it means the task schema will have a reference for the user schema.

(In the database, the tasks table will have a foreign key of users table)


Run:
`mix phx.gen.html Tasks Task tasks name:string description:text user_id:references:users`

In which:

- `mix phx.gen.html` is the mix task that generates controller, views,
and context for an HTML resource

- `Tasks` is the Phoenix Context to encapsulate the data access and
data validation for Tasks; yeah, naming things is hard :)
The suggestion from docs is to use the plural form of the resource when
we don’t have a clear context name

- `Tasks.Task` defines the Elixir struct name, which will be an Ecto
schema for casting and validating task data

- `tasks` is the database table name on PostgreSQL to store our task fields

- `name:string description:text` is the list of fields and its data type

- `user_id:references:users` defines the field name for the reference to 
the users table, in this case, it will be `user_id`


Important: after running the phoenix generator, it will suggest to
add a new resource on the `lib/taskify_web/router.ex` file and then, to run
`mix ecto.migrate`. Do not run it yet!

We need to make a few tweaks to the generated migration in
`priv/repo/migrations/*_create_tasks.exs` (See next commit)

Note: in the daily work, this commit and the next one should be a unique
commit; they are separated only for the workshop purpose




---

Context about Phoenix Contexts


In Phoenix, contexts often encapsulate data access and data validation.
Overall, think of them as boundaries to decouple and isolate parts of
your application.


About the context generators:

Using the context generators is a great way for beginners
and intermediate Elixir programmers alike to get up and running quickly
while thoughtfully designing their applications.


https://hexdocs.pm/phoenix/contexts.html
https://hexdocs.pm/phoenix/Mix.Tasks.Phx.Gen.Html.html